+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Add .gitattributes file
+
 ## [1.0.3] - 2025-08-11
 
 - Update to gg\_git 3.0.0
@@ -22,6 +28,7 @@
 
 - prepare version 1.0.1
 
+[Unreleased]: https://github.com/inlavigo/gg_merge/compare/1.0.3...HEAD
 [1.0.3]: https://github.com/inlavigo/gg_merge/compare/1.0.2...1.0.3
 [1.0.2]: https://github.com/inlavigo/gg_merge/compare/1.0.1...1.0.2
 [1.0.1]: https://github.com/inlavigo/gg_merge/tag/%tag